Починил позицию и ману в мире: координаты уходили с переставленной высотой, вход в мир был сдвинут на восемь байт, а максимумы HP и MP слались как int вместо short

+++ Packets/Packets.Server.Game/Models/Send/Character/Characteristics/5145_InventoryCharacteristicModel.cs
@@ -25,7 +25,15 @@ namespace Packets.Server.Game.Models.Send.Character.Characteristics
         public short Dex { get; set; }
         public short Int { get; set; }
         public short CriticalHit { get; set; }
-        public int HpMax { get; set; }
-        public int MpMax { get; set; }
+
+        /// <summary>
+        ///     AbInfo.mMaxHp of the original, a short
+        /// </summary>
+        public short HpMax { get; set; }
+
+        /// <summary>
+        ///     AbInfo.mMaxMp of the original, a short
+        /// </summary>
+        public short MpMax { get; set; }
     }
 }
